spacemacs-mac: enabled web-mode auto-closing/pairing
- see: http://web-mode.org/

Autopairing, tag autoclosing does not trigger

This is the normal behavior when you run emacs in a terminal (tty).
That’s because in text mode, pasting a long text is like inserting it
one character at a time. Thus, the consequences can be bad e.g. many
auto closing/pairing instead of one. If you want to force autoclosing,
autopairing even in a terminal, add

(setq web-mode-enable-auto-closing t)

and

(setq web-mode-enable-auto-pairing t)

in your .emacs